JumpClip/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jumpclip"
version = "0.1.0"
description = "Prompt-driven public-domain sprite sheet synthesis and analysis toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "Pillow>=10.3",
  "requests>=2.32",
]

[project.scripts]
jumpclip = "jumpclip.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
